requirements.txt
requests
python-daemon
pyyaml
websocket-client
slackclient
gspread
oauth2client
PyOpenSSL
